+++ CafeVesuviusApi/Controllers/OrdersController.cs
@@ -32,7 +32,7 @@ namespace CafeVesuviusApi.Controllers
 
         // GET: api/Orders/5
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetOrder(long id)
+        public async Task<IActionResult> GetOrder(int id)
         {
             if (await _orderRepository.GetOrders() == null) return NotFound();
             var order = await _orderRepository.GetOrder(id);
@@ -43,7 +43,7 @@ namespace CafeVesuviusApi.Controllers
         // PUT: api/Orders/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutOrder(long id, Order order)
+        public async Task<IActionResult> PutOrder(int id, Order order)
         {
             if (id != order.Id) return BadRequest();
             if(!(await _orderRepository.UpdateOrder(id, order))) return BadRequest();
@@ -62,7 +62,7 @@ namespace CafeVesuviusApi.Controllers
 
         // DELETE: api/Orders/5
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteOrder(long id)
+        public async Task<IActionResult> DeleteOrder(int id)
         {
             if (await _orderRepository.GetOrders() == null) return NotFound();
             if(!(await _orderRepository.DeleteOrder(id))) return BadRequest();
